Split DefaultLocalComponentMetadata into multiple classes
This PR splits DefaultLocalComponentMetadata into several top-level classes while attempting to
avoid any other non-essential changes. The intention is to pull out the non-static inner classes
representing the metadata of a Configuration and its Variants.
